Release 6.0.0
* Samples
  * Added ModelInstancing for GLES3 and Vulkan.
  * Added System.DevBasicCustomRender for GLES2, GLES3 and Vulkan.
  * Added UI.BasicDataBinding for GLES2, GLES3 and Vulkan.
  * Added UI.Charts for GLES2, GLES3 and Vulkan.
  * Converted SpatialHashGrid2D from Vulkan to GLES3.
  * UI.Benchmark added 'alternative' draw re-order algorithm.
* Libraries
  * FslBase replaced Optional with std::optional
  * New library FslDataBinding which provides basic data binding support.
  * FslGraphics3D.BasicRender: Added initial custom shader support.
  * FslSimpleUI: Added initial data binding support
  * FslSimpleUI.Charts:
    * Added BoxPlot control
    * Added initial data-binding support.
  * FslSimpleUI.Render cached font information for faster text rendering.
  * C++17 allowed a lot more things to be constexpr.
* Experimental Emscripten build support.
* Now require a compiler with C++17 support.
* Tool changes
  * **VS2022** is now the default on windows.
  * FslBuild tools added feature HW_GPU_VIVANTE for samples that require a vivante GPU.
  * The gtec-texture-packer used to create texture atlases for the samples is now freely available for everyone to use.
  * Now require at least CMake 3.16.
  * Now require clang format and clang tidy 13+
  * Android NDK r23c
* Misc:
  * Clang format was allowed to auto sort all includes.
  * Google Angle was re-tested to see if it could serve as the default OpenGL-ES emulator on windows.
  * Upgraded to Assimp 5.2.4.
  * Upgraded to FMT 8.1.1.
  * Upgraded to GoogleTest 1.11.0.
  * Upgraded to nlohmann/json 3.10.5.
  * Upgraded to RapidVulkan 1.3.211.
  * OpenVX was upgraded to 1.2

+++ DemoApps/GLES2/Blur/source/TestScene.cpp
@@ -29,13 +29,13 @@
  *
  ****************************************************************************************************************************************************/
 
+#include "TestScene.hpp"
 #include <FslBase/Log/Log3Fmt.hpp>
 #include <FslDemoApp/Base/Service/Content/IContentManager.hpp>
 #include <FslDemoService/Graphics/IGraphicsService.hpp>
 #include <FslGraphics/Bitmap/Bitmap.hpp>
 #include <FslGraphics/Vertices/VertexPositionTexture.hpp>
 #include "GausianHelper.hpp"
-#include "TestScene.hpp"
 
 
 namespace Fsl
@@ -76,8 +76,8 @@ namespace Fsl
 
     {    // prepare the vertex buffer
       // We scale the UV coordinates so that we get a 1-1 pixel mapping on the screen
-      const float scaleX = screenResolution.X / float(m_texture.GetSize().Width());
-      const float aspect = (screenResolution.Y / static_cast<float>(screenResolution.X));
+      const float scaleX = static_cast<float>(screenResolution.X) / static_cast<float>(m_texture.GetSize().Width());
+      const float aspect = (static_cast<float>(screenResolution.Y) / static_cast<float>(screenResolution.X));
       const float u1 = 0.0f;
       const float u2 = scaleX;
       const float v1 = scaleX * aspect;
