+++ DemoApps/OpenVG/Example3/source/Example3.cpp
@@ -29,12 +29,12 @@
  *
  ****************************************************************************************************************************************************/
 
+#include "Example3.hpp"
 #include <FslBase/Exceptions.hpp>
 #include <FslBase/Math/Vector2.hpp>
 #include <FslBase/UncheckedNumericCast.hpp>
 #include <FslUtil/OpenVG/VGCheck.hpp>
 #include <algorithm>
-#include "Example3.hpp"
 #include <array>
 #include <cmath>
 #include <iostream>
@@ -91,8 +91,8 @@ namespace Fsl
     FSLGRAPHICSOPENVG_CHECK_FOR_ERROR();
 
     const PxSize2D currentSizePx = GetWindowSizePx();
-    float scaleX = currentSizePx.Width() / 1280.0f;
-    float scaleY = currentSizePx.Height() / 1080.0f;
+    float scaleX = static_cast<float>(currentSizePx.Width()) / 1280.0f;
+    float scaleY = static_cast<float>(currentSizePx.Height()) / 1080.0f;
     std::vector<Vector2> vgTrianglePoints;
     std::vector<Vector2> vgQuadCurvePoints;
     std::vector<Vector2> vgCubicCurvePoints;
